@@ -36,21 +38,20 @@
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
 
-		<jms-api.version>2.0.1</jms-api.version>
 		<junit-jupiter.version>5.5.2</junit-jupiter.version>
-		<log4j2.version>2.13.3</log4j2.version>
+		<log4j.version>2.14.1</log4j.version>
 		<mockito.version>3.1.0</mockito.version>
-		<jackson.version>2.10.2</jackson.version>
+		<jackson.version>2.12.3</jackson.version>
 
-		<spring-framework.version>5.2.3.RELEASE</spring-framework.version>
-		<spring-boot.version>2.2.4.RELEASE</spring-boot.version>
+		<spring-framework.version>5.3.9</spring-framework.version>
+		<spring-boot.version>2.5.3</spring-boot.version>
 		<maven-compiler-plugin.version>3.8.1</maven-compiler-plugin.version>
 		<maven-gpg-plugin.version>1.6</maven-gpg-plugin.version>
 		<maven-javadoc-plugin.version>3.2.0</maven-javadoc-plugin.version>
 		<maven-source-plugin.version>3.2.0</maven-source-plugin.version>
 		<maven-surefire-plugin.version>2.22.2</maven-surefire-plugin.version>
 
-		<org.apache.activemq.version>5.15.8</org.apache.activemq.version>
+		<jms-api.version>2.0.1</jms-api.version>
 	</properties>
 
 	<dependencies>
@@ -209,9 +221,9 @@
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-compiler-plugin</artifactId>
-                <version>${maven-compiler-plugin.version}</version>
+				<version>${maven-compiler-plugin.version}</version>
 				<configuration>
-					<release>11</release>
+					<release>15</release>
 					<encoding>UTF-8</encoding>
 					<compilerArgs>
 						<arg>-parameters</arg>
