@@ -1,13 +1,13 @@
 package me.ehp246.aufjms.api.endpoint;
 
-import me.ehp246.aufjms.api.jms.Msg;
+import me.ehp246.aufjms.api.jms.JmsMsg;
 
 /**
  * 
  * @author Lei Yang
- *
+ * @since 1.0
  */
 @FunctionalInterface
 public interface ExecutableResolver {
-	ResolvedExecutable resolve(Msg msg);
+    Executable resolve(JmsMsg msg);
 }
