@@ -3,8 +3,8 @@ package me.ehp246.aufjms.api.endpoint;
 /**
  * 
  * @author Lei Yang
- *
+ * @since 1.0
  */
 public enum InstanceScope {
-	BEAN, MESSAGE,
+    BEAN, MESSAGE,
 }
