@@ -5,9 +5,9 @@ import javax.jms.Destination;
 /**
  * 
  * @author Lei Yang
- *
+ * @since 1.0
  */
 @FunctionalInterface
 public interface DestinationNameResolver {
-	Destination resolve(String name);
+    Destination resolve(String connectionName, String destinationName);
 }
