@@ -4,27 +4,27 @@ import java.util.List;
 import java.util.Map;
 
 public interface MsgSupplier {
-	String getCorrelationId();
+    String getCorrelationId();
 
-	String getType();
+    String getType();
 
-	String getInvoking();
+    String getInvoking();
 
-	List<?> getBodyValues();
+    List<?> getBodyValues();
 
-	default Long getTtl() {
-		return null;
-	}
+    default Long getTtl() {
+        return null;
+    }
 
-	default String getGroupId() {
-		return null;
-	}
+    default String getGroupId() {
+        return null;
+    }
 
-	default Map<String, String> getPropertyMap() {
-		return null;
-	}
+    default Map<String, String> getPropertyMap() {
+        return null;
+    }
 
-	default boolean isException() {
-		return false;
-	}
+    default boolean isException() {
+        return false;
+    }
 }
