@@ -2,21 +2,26 @@ package me.ehp246.aufjms.core.reflection;
 
 import java.util.concurrent.Callable;
 
+/**
+ * 
+ * @author Lei Yang
+ * @since 1.0
+ */
 public interface CatchingInvocation {
-	static <T> InvocationOutcome<T> invoke(final Callable<T> callable) {
-		try {
-			return InvocationOutcome.returned(callable.call());
-		} catch (final Exception e) {
-			return InvocationOutcome.thrown(e);
-		}
-	}
+    static <T> InvocationOutcome<T> invoke(final Callable<T> callable) {
+        try {
+            return InvocationOutcome.returned(callable.call());
+        } catch (final Exception e) {
+            return InvocationOutcome.thrown(e);
+        }
+    }
 
-	static InvocationOutcome<Void> invoke(final Runnable runnable) {
-		try {
-			runnable.run();
-			return InvocationOutcome.returned(null);
-		} catch (final Exception e) {
-			return InvocationOutcome.thrown(e);
-		}
-	}
+    static InvocationOutcome<Void> invoke(final Runnable runnable) {
+        try {
+            runnable.run();
+            return InvocationOutcome.returned(null);
+        } catch (final Exception e) {
+            return InvocationOutcome.thrown(e);
+        }
+    }
 }
