@@ -1,21 +1,21 @@
 package me.ehp246.aufjms.core.bymsg;
 
-import me.ehp246.aufjms.api.annotation.ByMsg;
+import me.ehp246.aufjms.api.annotation.ByJms;
 
 public interface DestinationTestCase {
 
-	@ByMsg(value = "test.inbox")
-	interface Case001 {
-		void m001();
-	}
+    @ByJms(destination = "test.inbox")
+    interface Case001 {
+        void m001();
+    }
 
-	@ByMsg("")
-	interface Case002 {
-		void m001();
-	}
+    @ByJms(destination = "")
+    interface Case002 {
+        void m001();
+    }
 
-	@ByMsg(value = "test.inbox")
-	interface Case003 {
-		void m001();
-	}
+    @ByJms(destination = "test.inbox")
+    interface Case003 {
+        void m001();
+    }
 }
