@@ -2,9 +2,9 @@ package me.ehp246.aufjms.global.case002;
 
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-import me.ehp246.aufjms.api.annotation.EnableByMsg;
-import me.ehp246.aufjms.api.annotation.EnableForMsg;
-import me.ehp246.aufjms.api.annotation.EnableForMsg.At;
+import me.ehp246.aufjms.api.annotation.EnableByJms;
+import me.ehp246.aufjms.api.annotation.EnableForJms;
+import me.ehp246.aufjms.api.annotation.EnableForJms.At;
 import me.ehp246.aufjms.global.case002.request.Calculator;
 
 /**
@@ -12,7 +12,7 @@ import me.ehp246.aufjms.global.case002.request.Calculator;
  *
  */
 @SpringBootApplication
-@EnableByMsg
-@EnableForMsg(@At(value = "queue://me.ehp246.aufjms.request", scan = Calculator.class))
+@EnableByJms
+@EnableForJms(@At(value = "queue://me.ehp246.aufjms.request", scan = Calculator.class))
 class AppConfiguration {
 }
