@@ -4,7 +4,7 @@ import java.util.concurrent.atomic.AtomicReference;
 
 import org.springframework.stereotype.Service;
 
-import me.ehp246.aufjms.api.annotation.ForMsg;
+import me.ehp246.aufjms.api.annotation.ForJms;
 import me.ehp246.aufjms.api.annotation.Invoking;
 import me.ehp246.aufjms.api.endpoint.InstanceScope;
 
@@ -13,28 +13,28 @@ import me.ehp246.aufjms.api.endpoint.InstanceScope;
  *
  */
 @Service
-@ForMsg(scope = InstanceScope.BEAN)
+@ForJms(scope = InstanceScope.BEAN)
 public class Calculator {
-	public AtomicReference<Integer> mem = new AtomicReference<>();
+    public AtomicReference<Integer> mem = new AtomicReference<>();
 
-	@Invoking("setMem")
-	public void setMem(int i) {
-		this.mem.set(i);
-	}
+    @Invoking("setMem")
+    public void setMem(int i) {
+        this.mem.set(i);
+    }
 
-	@Invoking("addMem")
-	public int addMem(int i) {
-		this.mem.set(this.mem.get() + i);
-		return mem.get();
-	}
+    @Invoking("addMem")
+    public int addMem(int i) {
+        this.mem.set(this.mem.get() + i);
+        return mem.get();
+    }
 
-	@Invoking("getMem")
-	public int getMem() {
-		return this.mem.get();
-	}
+    @Invoking("getMem")
+    public int getMem() {
+        return this.mem.get();
+    }
 
-	@Invoking
-	public int add(int i, int j) {
-		return i + j;
-	}
+    @Invoking
+    public int add(int i, int j) {
+        return i + j;
+    }
 }
