@@ -8,7 +8,7 @@ import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 
 import me.ehp246.aufjms.api.annotation.CollectionOf;
-import me.ehp246.aufjms.api.annotation.ForMsg;
+import me.ehp246.aufjms.api.annotation.ForJms;
 import me.ehp246.aufjms.api.annotation.Invoking;
 import me.ehp246.aufjms.api.endpoint.InstanceScope;
 import me.ehp246.aufjms.api.endpoint.InvocationModel;
@@ -18,36 +18,36 @@ import me.ehp246.aufjms.api.endpoint.InvocationModel;
  *
  */
 @Service
-@ForMsg(scope = InstanceScope.BEAN, invocation = InvocationModel.SYNC)
+@ForJms(scope = InstanceScope.BEAN, invocation = InvocationModel.SYNC)
 class Alarm {
-	private Instant[] instants;
+    private Instant[] instants;
 
-	@Invoking("setArray")
-	public void set(final Instant... instants) {
-		this.instants = instants;
-	}
+    @Invoking("setArray")
+    public void set(final Instant... instants) {
+        this.instants = instants;
+    }
 
-	@Invoking("setCollection")
-	public void set(@CollectionOf(Instant.class) final Set<Instant> instants) {
-		// Type checking
-		this.instants = instants.toArray(new Instant[] {});
-	}
+    @Invoking("setCollection")
+    public void set(@CollectionOf(Instant.class) final Set<Instant> instants) {
+        // Type checking
+        this.instants = instants.toArray(new Instant[] {});
+    }
 
-	@Invoking("get")
-	public List<Instant> get() {
-		return List.of(instants);
-	}
+    @Invoking("get")
+    public List<Instant> get() {
+        return List.of(instants);
+    }
 
-	@Invoking("flatSet")
-	public List<Set<List<Instant>>> flatSet(
-			@CollectionOf({ Set.class, List.class, Instant.class }) final List<Set<List<Instant>>> instants) {
-		final var list = instants.stream().flatMap(Set::stream).flatMap(List::stream)
-				.collect(Collectors.<Instant>toList());
-		this.instants = new Instant[list.size()];
+    @Invoking("flatSet")
+    public List<Set<List<Instant>>> flatSet(
+            @CollectionOf({ Set.class, List.class, Instant.class }) final List<Set<List<Instant>>> instants) {
+        final var list = instants.stream().flatMap(Set::stream).flatMap(List::stream)
+                .collect(Collectors.<Instant>toList());
+        this.instants = new Instant[list.size()];
 
-		for (int i = 0; i < list.size(); i++) {
-			this.instants[i] = list.get(i);
-		}
-		return instants;
-	}
+        for (int i = 0; i < list.size(); i++) {
+            this.instants[i] = list.get(i);
+        }
+        return instants;
+    }
 }
