@@ -2,16 +2,16 @@ package me.ehp246.aufjms.global.destination.case001;
 
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-import me.ehp246.aufjms.api.annotation.EnableByMsg;
-import me.ehp246.aufjms.api.annotation.EnableForMsg;
-import me.ehp246.aufjms.api.annotation.EnableForMsg.At;
+import me.ehp246.aufjms.api.annotation.EnableByJms;
+import me.ehp246.aufjms.api.annotation.EnableForJms;
+import me.ehp246.aufjms.api.annotation.EnableForJms.At;
 
 /**
  * @author Lei Yang
  *
  */
 @SpringBootApplication
-@EnableByMsg(replyTo = "topic://${aufjms.reply.topic}")
-@EnableForMsg({ @At("queue://${aufjms.request.queue}"), @At("${aufjms.request.queue02}") })
+@EnableByJms(replyTo = "topic://${aufjms.reply.topic}")
+@EnableForJms({ @At("queue://${aufjms.request.queue}"), @At("${aufjms.request.queue02}") })
 public class DestinationConfiguration001 {
 }
