@@ -2,16 +2,16 @@ package me.ehp246.aufjms.global.destination.case002;
 
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-import me.ehp246.aufjms.api.annotation.EnableByMsg;
-import me.ehp246.aufjms.api.annotation.EnableForMsg;
+import me.ehp246.aufjms.api.annotation.EnableByJms;
+import me.ehp246.aufjms.api.annotation.EnableForJms;
 
 /**
  * @author Lei Yang
  *
  */
 @SpringBootApplication
-@EnableByMsg
-@EnableForMsg
+@EnableByJms
+@EnableForJms
 public class DestinationConfiguration002 {
 
 }
