@@ -8,32 +8,32 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 
-import me.ehp246.aufjms.api.endpoint.MsgEndpoint;
+import me.ehp246.aufjms.api.endpoint.AtEndpoint;
 
 /**
  * @author Lei Yang
  *
  */
 @SpringBootTest(classes = DestinationConfiguration002.class, properties = {
-		"spring.activemq.broker-url=vm://activemq?broker.persistent=false&broker.useShutdownHook=false" })
+        "spring.activemq.broker-url=vm://activemq?broker.persistent=false&broker.useShutdownHook=false" })
 public class DestinationTest002 {
-	@Autowired
-	private AnnotationConfigApplicationContext appCtx;
+    @Autowired
+    private AnnotationConfigApplicationContext appCtx;
 
-	@Test
-	void destination001() {
-		final var endpoints = appCtx.getBeansOfType(MsgEndpoint.class);
+    @Test
+    void destination001() {
+        final var endpoints = appCtx.getBeansOfType(AtEndpoint.class);
 
-		Assertions.assertEquals(2, endpoints.size());
+        Assertions.assertEquals(2, endpoints.size());
 
-		final var names = endpoints.values().stream().map(MsgEndpoint::getDestinationName).collect(Collectors.toSet());
+        final var names = endpoints.values().stream().map(AtEndpoint::getDestinationName).collect(Collectors.toSet());
 
-		Assertions.assertEquals(true,
-				names.contains("queue://" + DestinationConfiguration002.class.getName() + ".request"),
-				"Should have the default request queue name");
+        Assertions.assertEquals(true,
+                names.contains("queue://" + DestinationConfiguration002.class.getName() + ".request"),
+                "Should have the default request queue name");
 
-		Assertions.assertEquals(true,
-				names.contains("topic://" + DestinationConfiguration002.class.getName() + ".reply"),
-				"Should have the default reply topic name");
-	}
+        Assertions.assertEquals(true,
+                names.contains("topic://" + DestinationConfiguration002.class.getName() + ".reply"),
+                "Should have the default reply topic name");
+    }
 }
